pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "graphviz>=0.21",
    "ipython>=8.37.0",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
]
